+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * 【all】        部分接口添加FunctionalInterface修饰
 * 【crypto】     KeyUtil增加readKeyStore重载
 * 【extra】      JschUtil增加私钥传入支持（issue#INKDR@Gitee）
+* 【core】       DateUtil、DateTime、Convert全面支持jdk8的time包
 
 ### Bug修复
 * 【http】       修复Cookie中host失效导致的问题（issue#583@Github）